seperated functions into middleware as test
